move IsStringPresent to util package
